requirements.txt
streamlit>=1.0
pandas>=1.5
mysql-connector-python>=8.0
python-dotenv>=1.2.1
